requirements.txt
pytesseract
pillow
torch 
torchvision 
torchaudio
easyocr